Ajustes na publicação. Movi as publicações para um outro pacote para separar do usuário.
