+++ backend/src/main/java/br/com/projetoestudos/projetoestudos/publication/DTO/PublicationDTO.java
@@ -1,0 +1,4 @@
+package br.com.projetoestudos.projetoestudos.publication.DTO;
+
+public class PublicationDTO {
+}
