--- backend/src/main/java/br/com/projetoestudos/projetoestudos/user/DTO/PublicationDTO.java
@@ -1,4 +1,0 @@
-package br.com.projetoestudos.projetoestudos.user.DTO;
-
-public class PublicationDTO {
-}
