--- backend/src/main/java/br/com/projetoestudos/projetoestudos/user/controller/PublicationController.java
@@ -1,4 +1,0 @@
-package br.com.projetoestudos.projetoestudos.user.controller;
-
-public class PublicationController {
-}
