+++ backend/src/main/java/br/com/projetoestudos/projetoestudos/user/model/User.java
@@ -1,9 +1,12 @@
 package br.com.projetoestudos.projetoestudos.user.model;
 
+import br.com.projetoestudos.projetoestudos.publication.model.Publication;
 import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
+
+import java.util.List;
 
 @Entity(name = "user")
 @Getter
@@ -15,4 +18,8 @@ public class User {
 
     @Column(name = "name", length = 60)
     private String name;
+
+    @JoinColumn(name = "publications")
+    @OneToMany(fetch = FetchType.LAZY)
+    private List<Publication> publications;
 }
